Upload Box Functionality
Added some visual indication that you have selected some files from your computer.

Added the abililty to add files to the list of uploading files when you drag-and-drop them from a file explorer.

Realized after all this that we probably don't want to multi-upload for a singular publication upload. Going to have to go back and figure out how to adapt what we've done to a single file.